fix(wta): pin /model & /-autocomplete popups directly above the input box
Both input-anchored popups were positioned relative to the filler row
(`chunks[3]`, the `Min(0)` gap between the perm panel and the hint row)
instead of the input box. The popup anchors to the *top* of whatever rect
it's given, so a short chat — which leaves a tall filler — floated the
popup far up the pane. Worst in side-by-side layouts where each pane holds
less content. Anchor to the input box (`chunks[6]`) so the popup's bottom
edge always lands on the input's top edge, fixed regardless of empty space
above.

Also dedup the shared popup chrome into `ui/popup.rs`: the above/below
positioning algorithm (`anchored_above`), the bordered+titled block
(`block`), and the `BORDER_HEIGHT` constant — previously copy-pasted across
model_popup.rs and command_popup.rs. The centered /help overlay reuses
`block` for its frame. Per-popup bits (max-visible counts, list contents,
/help centering) stay local.
